==> processing_pipelines/Xenium_mouse_replicate_1_processing/10_Calculating_Spatial_Decomposition.ipynb <==
import tensorflow as tf
import scanpy as sc
import os
from scipy.spatial import KDTree
from tensorflow import keras
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.decomposition import NMF
from tqdm.notebook import tqdm
from core_functions.neighborhood_decomposition import *
import glob

### Spatial decomposition is calculatedto use as a feature set in defining axes. We perform spatial decomposition in the following script

output_folder = r"D:/amonell/timecourse_final/analysis/cleaned"

input_folders = glob.glob("D:/amonell/timecourse_final/day*")

combined_adata = sc.read(os.path.join(output_folder, "integrated_celltyped.h5ad"))

Celltype classes that should remain relatively consistent at different timepoints pi. Since we calculate the dimensionality reduction on all datasets simulataneously, we expect to capture spatially conerved features across timepoints rather than variation caused by a changing infection.

unchanging_type_keys = ["Epithelial", "Stromal"]

combined_adata_no_immune = combined_adata[
    combined_adata.obs["Class"].isin(unchanging_type_keys)
]

unique_batches = np.unique(combined_adata_no_immune.obs.new_batch.values)

nneighbors = 10
dfs = []
for input_file in input_folders:
    adata = combined_adata_no_immune[
        combined_adata_no_immune.obs["batch"] == os.path.basename(input_file)
    ]
    adata_arr = np.array(adata.X)
    celltype_cluster = adata.obs.index.values
    list_of_arrays = []
    spatial_points = np.array(
        [adata.obsm["X_spatial"][:, 0], adata.obsm["X_spatial"][:, 1]]
    ).T
    tree = KDTree(spatial_points)
    for i_bac in tqdm(range(len(celltype_cluster))):
        current_cell = celltype_cluster[i_bac]
        distances, neighbors = tree.query(spatial_points[i_bac], k=nneighbors)
        neighbors = np.array(list(neighbors))
        gene_array = np.array(np.sum(adata_arr[neighbors, :], axis=0)).squeeze()
        list_of_arrays.append(gene_array)

    X = pd.DataFrame(np.array(list_of_arrays))
    dfs.append(X)

X_arr = pd.concat(dfs)

num_neighborhoods = 15
X = X_arr
f = len(X.columns)
n = len(X.index.tolist())

model = NMF(n_components=num_neighborhoods, random_state=0)
W = model.fit_transform(X)
H = model.components_

for input_file in input_folders:
    if os.path.exists(os.path.join(input_file, "adatas", "05_reference_unrolled.h5ad")):
        adata = sc.read(
            os.path.join(input_file, "adatas", "05_reference_unrolled.h5ad")
        )
    else:
        adata = sc.read(os.path.join(input_file, "adatas", "05_unrolled.h5ad"))

    superclusters = (
        combined_adata[combined_adata.obs["batch"] == os.path.basename(input_file)]
        .obs["Class"]
        .values
    )
    cluster_observations = combined_adata[
        combined_adata.obs["batch"] == os.path.basename(input_file)
    ].obs[["leiden", "Sub_leiden", "Class", "Type", "Subtype", "Immunocentric_Type"]]
    cluster_observations.index = [
        i.split("-")[0] for i in cluster_observations.index.values
    ]
    adata.obs = adata.obs.merge(
        cluster_observations, left_on="cell", right_index=True, how="left"
    )
    celltype_cluster = adata.obs.index.values
    base_dictionary = {}
    for i in np.unique(celltype_cluster):
        base_dictionary[i] = 0
    nneighbors = 10
    list_of_arrays = []
    adata_epi = adata[adata.obs["Class"].isin(unchanging_type_keys)]
    spatial_points_epi = np.array(
        [adata_epi.obsm["X_spatial"][:, 0], adata_epi.obsm["X_spatial"][:, 1]]
    ).T
    spatial_points = np.array(
        [adata.obsm["X_spatial"][:, 0], adata.obsm["X_spatial"][:, 1]]
    ).T
    adata_epi_arr = np.array(adata_epi.X)

    tree = KDTree(spatial_points_epi)
    for i_bac in tqdm(range(len(celltype_cluster))):
        current_cell = celltype_cluster[i_bac]
        distances, neighbors = tree.query(spatial_points[i_bac], k=nneighbors)
        neighbors = np.array(list(neighbors))
        gene_array = np.array(np.sum(adata_epi_arr[neighbors, :], axis=0)).squeeze()
        list_of_arrays.append(gene_array)

    X = pd.DataFrame(np.array(list_of_arrays)).astype(H.dtype)
    W = model.transform(X)

    topics_frame = pd.DataFrame(W)

    topics_frame.columns = [
        "Topic " + str(i + 1) for i in range(len(topics_frame.columns))
    ]
    topics_frame.index = adata.obs.index.tolist()

    def zscore(column):
        return (column - column.mean()) / column.std()

    # Apply the z-score function to each column in the dataframe
    topics_frame = topics_frame.apply(zscore)
    adata.obs = adata.obs.merge(topics_frame, left_index=True, right_index=True)
    adata.obs["topic"] = pd.Categorical(
        (np.argmax(topics_frame.values, axis=1) + 1).astype(str)
    )

    sc.set_figure_params(dpi=300)
    figure = sc.pl.embedding(
        adata,
        basis="spatial",
        color="topic",
        vmax=1,
        cmap="Blues",
        title="Neighborhood",
        size=2,
        show=False,
        return_fig=True,
    )
    try:
        os.mkdir(os.path.join(input_file, "figures", "neighborhoods"))
    except:
        print("Figures/neighborhoods already made.")
    figure.tight_layout()
    plt.axis("equal")
    figure.savefig(
        os.path.join(input_file, "figures", "neighborhoods", "neighborhoods.png")
    )
    plt.close()
    adata.write(
        os.path.join(input_file, "adatas", "06_before_decomposition_model.h5ad")
    )
